Install Euphonic to Mantid-visible Python using Pip.
This is intended as a temporary stop-gap for users importing force constant data to SimulatedDensityOfStates or Abins, while using a version of Mantid that was not built with Conda. (Conda-built Mantid should come with Euphonic pre-installed.)

The script may start by updating pip to get consistent behaviour. The main thing the script is trying not to do is update Numpy, as a Pip-installed Numpy might not be binary-compatible with Mantid. To avoid changing the user python setup more than necessary, some auxiliary tools are installed to a temporary directory. Euphonic is installed to the visible user-install location, which is probably something like ~/.local/lib/python3/site-packages.